Rename placeholders.tsx to stencil-placeholders.tsx
https://coveord.atlassian.net/browse/KIT-4199
